backend/requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
requests
